Fix: unexpected combat when FLEET_2_STEP_ON = True in 7-2
The scale icon on the top-left of enemy at A3 might be covered by the bumping arrow above current fleet at A5 when camera is focusing to D3 or D2 at beginning. Then the enemy at A3 cannot be recognized, and result in an unexpected combat if FLEET_2_STEP_ON = True.
Adding A2 to camera_data_spawn_point can fix this problem. It only needs to move camera horizontally when scanning A2 after D2, so A2 (instead of A3 or A1) is added.